api/predictGemstone/requirements.txt
requests
opencv-python-headless
Pillow
numpy
pandas
tensorflow==2.7.0